+++ src/__tests__/suite/noFolder.test.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
+import { suite, after, before } from 'mocha';
 
 import { activateExtension } from './helpers';